requirements.txt
# Core dependencies
jinja2>=3.1.0
pyyaml>=6.0
requests>=2.28.0

# AI/LLM
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# Environment
python-dotenv>=1.0.0

# Optional: for advanced features
beautifulsoup4>=4.11.0
markdown>=3.4.0


